requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
lightgbm>=4.0
matplotlib>=3.7
Pillow>=10.0
